requirements.txt
discord.py~=2.3.2
cryptography~=42.0.8

requests~=2.31.0
pillow~=10.3.0